# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: Azile AI MDM API
  version: "3.1.0"
  description: |
    Azile AI MDM is a hub-and-spoke Master Data Management platform built on Rust microservices.

    ## Services
    | Service | Port | Purpose |
    |---------|------|---------|
    | API Gateway | 8080 | Single entry point; auth + rate-limit + proxy |
    | MDM Core | 8081 | Entities, matching, merge, approval, admin config |
    | AI Service | 8082 | LLM copilot, embeddings, semantic matching, RAG |
    | Policy Service | 8083 | Policy evaluation, GDPR, consent |
    | Ingest Service | 8084 | Batch/CSV/REST ingestion, source systems |
    | Search Service | 8085 | Full-text search, autocomplete |
    | Notification Service | 8086 | In-app notifications, webhooks, email |

    ## Authentication
    All protected endpoints require a JWT Bearer token obtained from `POST /auth/login`.
    SCIM 2.0 endpoints use a separate SCIM bearer token obtained from `POST /scim/tokens`.

    Standard request headers:
    - `Authorization: Bearer <jwt>` â€” JWT from login
    - `x-tenant-id` â€” Tenant UUID (injected by gateway; required on direct calls)
    - `x-correlation-id` â€” Optional idempotency / tracing ID

  contact:
    name: Azile MDM Support
    email: ajit.more@lithytree.com
  license:
    name: Proprietary
servers:
  - url: http://localhost:8080
    description: API Gateway (recommended entry point)
  - url: http://localhost:8081
    description: MDM Core (direct)
...
security:
  - bearerAuth: []
...
paths:
...
  /mcp/query:
    post:
      tags: [AI Copilot]
      summary: Send a natural-language query to the AI Copilot
      servers:
        - url: http://localhost:8082
          description: AI Service
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/McpRequest" }
      responses:
        "200":
          description: Copilot response
          content:
            application/json:
              schema: { $ref: "#/components/schemas/McpResponse" }
  /mcp/stream:
    post:
      tags: [AI Copilot]
      summary: Streaming copilot query (Server-Sent Events)
      servers:
        - url: http://localhost:8082
          description: AI Service
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/McpRequest" }
      responses:
        "200":
          description: SSE stream of token chunks
          content:
            text/event-stream:
              schema:
                type: string
                description: 'Lines of the form: data: {"chunk": "token"}'
  /match/semantic:
    post:
      tags: [Matching]
      summary: AI-powered semantic match scoring
      servers:
        - url: http://localhost:8082
          description: AI Service
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/SemanticMatchRequest" }
      responses:
        "200":
          description: Semantic similarity score
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }
  /embed:
    post:
      tags: [AI Copilot]
      summary: Generate a text embedding vector
      servers:
        - url: http://localhost:8082
          description: AI Service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string }
      responses:
        "200":
          description: Embedding vector
          content:
            application/json:
              schema:
                type: object
                properties:
                  success: { type: boolean }
                  embedding:
                    type: array
                    items: { type: number, format: float }
                  dims: { type: integer }
  /rag/index:
    post:
      tags: [AI Copilot]
      summary: Index a document into the RAG knowledge base
      servers:
        - url: http://localhost:8082
          description: AI Service
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [tenant_id, doc_type, title, content]
              properties:
                tenant_id: { type: string, format: uuid }
                doc_type: { type: string }
                title: { type: string }
                content: { type: string }
      responses:
        "201":
          description: Document indexed
  /weights/recommend:
    get:
      tags: [Matching]
      summary: AI-recommended matching field weights based on steward feedback
      servers:
        - url: http://localhost:8082
          description: AI Service
      parameters:
        - name: tenant_id
          in: query
          required: true
          schema: { type: string, format: uuid }
        - name: min_samples
          in: query
          schema: { type: integer, default: 50 }
      responses:
        "200":
          description: Weight recommendations
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiResponse" }

  # â”€â”€ Ingest Service â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT obtained from POST /auth/login

    scimBearerAuth:
      type: http
      scheme: bearer
      bearerFormat: SCIM Token
      description: SCIM bearer token obtained from POST /scim/tokens
...
  schemas:
...
    ApiResponse:
      type: object
      properties:
        success: { type: boolean }
        data: {}
        message: { type: string }
        error: { type: string, nullable: true }
...
    SemanticMatchRequest:
      type: object
      required: [tenant_id, source_attrs, candidate_attrs, entity_type]
      properties:
        tenant_id: { type: string, format: uuid }
        source_attrs: { type: object, additionalProperties: true }
        candidate_attrs: { type: object, additionalProperties: true }
        algo_score: { type: number, format: float }
        entity_type: { type: string }

    # â”€â”€ Merge â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
...
    McpRequest:
      type: object
      required: [tenant_id, user_id, prompt]
      properties:
        tenant_id: { type: string, format: uuid }
        user_id: { type: string, format: uuid }
        prompt: { type: string }
        tool: { type: string, nullable: true }
        response_format: { type: string, enum: [text, json], default: text }
    McpResponse:
      type: object
      properties:
        success: { type: boolean }
        result: { type: string }
        tool_results: { type: array, items: { type: object } }
        error: { type: string, nullable: true }

    # â”€â”€ Ingestion â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€â”€
...